Fix rnnet not reusing created inference sessions (#784)
## Summary
This fixes an issue where we created a new instance of the RNNet model
every time we perform a fold instead of instantiating it once the first
time it's run and then reusing it (causing runaway memory usage and
adding additional unnecessary computation overhead).

## Implementation Notes
We changed to "memoizing" the inference session so that we could only
load rnnet when required instead of always loading it when the app
starts. However, I neglected to actually save the value after
initializing it, so it was always recreated (plus it appears that due to
the implementation of InferenceSession, its resources are not released
when it goes out of scope).