Avoid surgical updates of undo state when setting custom target structure from solution (#809)
Previously, we set the target structure for a solution after an
undoblock was created. This was part historical artifact (eg, due to
using pasteSequence to update the sequence which wound up creating a new
undoblock, which was changed in #768). That said, I also didn't realize
there was a better way: set `_targetPairs`/`_targetOligosOrder`, like
what happens when magic glue is used - these are intended to be
"scratch" variables that are updated with changes and propagated to the
undoblock once it's created. This is simpler, more idiomatic, and avoids
any special handling needed if we're in "freeze mode"