Display error with ID of solution if error is encountered when loading solution list (#810)
## Summary
We've recently run into situations where errors are thrown when parsing
solutions, eg due to an invalid estimate structure. This currently
requires a fair amount of extra work to track down which solution is
causing issues. Now, we explicitly include the ID of the solution that
encountered an error in the error message.

## Testing
Validated when loading design browser for a puzzle which had a solution
with an invalid estimate structure