Dont show custom target warning if the target structure hasn't been changed from the default (#814)
## Summary
Currently, we have a warning message that displays if a region with an
unconstrained target structure has target pairs that are not filled with
bases that make valid pairs. This was to prevent the situation where a
custom target structure is created, but is later modified (eg, by
copying someone else's solution) while ignoring that target (ultimately,
we want to avoid saving a bogus custom target structure).

However, this warning was also showing if the default target structure
was left unchanged and the player elects to ignore it. Eg, in the
OpenVaccine puzzles, we do provide a default target structure which is
just a copy of the natural mode structure. The goal is to optimize
degredation, which likely requires a different structure, and players
should ignore the target entirely unless they want to define their own
target. It's fine that the target structure is recorded as the default
target structure of the puzzle - while it is "bogus" in a sense, we can
verify that it was indeed just the starting target, and we don't want to
provide a notice to the user where it's both confusing (they didn't
define a custom target, why is it complaining about it? also they may
not even know this feature exists, as it's generally an advanced
feature) and leading to unnecessary friction (extra clicks on submit).

## Testing
Tried submitting valid and invalid sequences for default and modified
target