Prevent HTML UI interaction as part of UI/dialog locks (#817)
## Summary
There are a number of situations when we "lock" the UI to user
interaction, such as when a modal dialog is present or we're in the
middle of folding. However, we weren't preventing interaction with the
parts of the UI in the HTML overlay. One way this manifested was if you
hovered over the script constraint codon chart while folding occurred,
an error is thrown complaining about trying to use a getter/setter (in
this case, get_sequence_string) while in the middle of folding.

## Implementation Notes
The early handler shenanigans are borrowed from ScrollContainer. It
might be a good idea at some point to move this stuff to some shared
utility, or find a way around it all together, but I don't have the
bandwidth to do that right now. This could be further improved if we
could also apply "dimming" to the HTML UI, but that also would require a
bit more work since we only want to dim the portions of the UI below the
HTML overlay, but not double-dim the canvas - and not apply a dim to
things drawn on the canvas that are supposed to be over the overlay.
Likely requires some CSS filter shenanigans?

## Testing
Normal interaction still works, interaction with the codon chart with a
UI lock is prevented, can still dismiss a notification dialog with a UI
lock.