Fix energy_of_structure_async not returning a promise (#828)
## Summary
Previously, the `energy_of_structure_async` Eternascript function was
identical to `energy_of_structure`. Now, it is an async function
returning a promise, as intended and described

## Implementation Notes
While no folder currently exposes scoreStructures as async, which is why
this did not lead to a type error, and there is nothing for us to await,
we are exposing this async version to allow users to future-proof for
situations where async access may be required.