Fix solutions being reloaded when switching to feedback mode (#831)
## Summary
Previously, when switching to feedback mode we re-queried solutions,
even if we already passed them from the design browser.
Also, even if it was required, it was being loaded *twice*. This changes
the behavior to use the previously-fetched
solutions if available without refetching, and if a fetch is needed,
only fetching once. This is particularly valuable for
puzzles with large amounts of solutions where it is time and memory
intensive to load all solutions.

## Testing
Loaded feedback view for OKR2 via design browser and directly