Allow tloop to be formed with sequence regions in any order (#837)
Discussed with Rhiju. Originally I had the impression that one of the
sequences should always be before the other, but they can actually be in
either order. The stamper already inadvertently stamped the structure in
both orders, but now the constraint also allows it (and the UI + code is
updated to refer to the two regions as a and b rather than 5' and 3')