Refactor/extract scripting api 2 (#840)
## Summary
This is part of the process of extracting the script-facing API from
PoseEditMode in order to make it usable for non-booster scripts.
Replaces #749

## Implementation Notes
I did it like the previous PR, except I didn't implement
get_folder/select_folder, since they are behind a "lock", and I didn't
want to think yet of whether or not I want the extracted logic to
inherit the lock.
However, while get_folder is locked, the calculation functions (such as
`fold`) which also access the folder, aren't locked (even on `dev`). Is
this intended?

## Testing
I ran https://eternagame.org/scripts/13774466, and it passed.
The script checks that all of the extracted functions (expect folding
with binding sites) work and return the same values.

This is an unrelated issue, but : For some reason, both on main/master
(https://eternagame.org/puzzles/6096060/play) and on my branch, the
first time I run it I get the following errors, and on future runs it
passes. Reloading the page makes the issue appear again for a single
execution.
Errors:

![image](https://github.com/user-attachments/assets/1fca8f90-85f4-485c-9ff4-a44fcdb2445a)
Successful run:

![image](https://github.com/user-attachments/assets/ecfe8020-2724-44f3-819f-0a6c13182786)
(In my branch I just copy pasted the script, since I didn't have it on
eternadev.org, and changed `Lib.fold`->`applet.fold`. I shouldn't have
Lib in the terminal, right?)


## Related Issues
Enables https://github.com/eternagame/eternagame.org/issues/378,
https://github.com/eternagame/EternaJS/pull/750