Dont allow p(unpaired) to be less than 0 (#843)
Conventionally, dot plots are probability matrices such that for any
base, the probability that it will pair with any other base will sum to
1, less the probability that it is unpaired - with probability unpaired
excluded, so to get the probability unpaired we sum and subtract from 1.
However for an algorithm like ribonanzanet, the dot plot is instead a
likelihood matrix such that the liklihood of each pair is independent of
every other pair, meaning it could sum to more than 1. We now clip
P(unp) to so that if the row sums to a value > 1, we don't drop p(unp)
to be below 0, which doesn't make sense (it is still ambiguous what
p(unp) should be if you have two pairs at .75 liklihood, but this is at
least _better_).

This reflects a change made in https://github.com/DasLab/arnie/pull/50