Support user event monitoring (#852)
## Summary
This provides a mechanism for monitoring user behavior, which could be
enabled in scenarios such as user studies which are researching
effectiveness of new tools or UI changes. This is specifically motivated
by an upcoming study on integration of automated solving tools.

## Implementation Notes
* The actions I've instrumented are not necessarily comprehensive, but
cover most of the key puzzle solving features. We can always add to this
over time/as we identify things we want to instrument
* The "namespacing" of events is hopefully convenient for filtering
in/out events that we may want to track in different scenarios (eg
specific user tests, or even monitoring via Grafana/Mimir), but this may
need tweaking
* I've restored move tracking and merged it with this system under the
`Move` namespace. The logic is largely unchanged from what was shown in
commit f06f1727984b8c951672f6e455de6c879af683e2, with the exception of
the reporting mechanism and paste having been moved into GameMode (we
now treat this as a paste move rather than a set of muations!)
* The ObservabilityReporter abstraction admittedly looks a little
excessive right now, but there was some rationale here. With this in
place, it becomes trivial to have two different sets of reports, each of
differing lifetimes, filtering a different set of events, going to
different places. For example, For this user study, we want user
interaction and moveset events for a given puzzle to be sent via
postmessage to Qualtrics, at some different point we may want to, say,
send all script execution events to Mimir over the entire app lifetime
(including as the puzzle changes) to have an indication of what scripts
and what script APIs are actively in use to ensure compatibility, and we
could even use this same mechanism to store movesets like we used to for
an individual puzzle.
* I made a couple of drive-by cleanups: a couple missing signal
registrations in PoseEdit, and tying the download keyboard shortcut into
the toolbar button

## Testing
* Some basic unit tests
* Enabled ConsoleReporter, clicked a couple buttons (mutated bases, swap
in hotbar and extended toolbar, change folder and marker layer)