Reverse complement target condition (#858)
## Summary
Adds support for a condition which makes the sequence the reverse
complement of other states. This is required for upcoming puzzles where
we wish to model both side-by-side as separate states, to gain insight
into transcription dynamics.

## Implementation Notes
The transformSequence mechanics used for an early FMN simulation
conveniently does something similar to what we need. However we also
needed to introduce this mechanism on pose edit and when switching
states in natural mode (bugs with this unused functionality).

## Testing
Test puzzle. Swapped between natural/target and states in pip and
non-pip. Made mutations in pip and non-pip and verified they translated
properly