Make mutation markers the default layer when a mutation constraint is active (#863)
## Summary
If a script constraint adds a custom marker layer, that marker will be
switched to be the default when its added. However, we did not do the
same for mutations. We now do so in the case a mutation constraint is
present

## Implementation Notes
This was discussed in a team meeting and determined to be desirable
behavior. If no mutation constraint is active, tracking mutations might
be interesting (hence we still add the layer) but it is not an intended
mechanic of the puzzle. However because the user actively needs to keep
track of this when a mutation constraint is present, it makes sense to
have this be the default mode.

## Testing
Tested puzzle with and without mutation constraint, does or does not set
the default layer as desired