Reposition barcode region in reverse-complement states (#864)
## Summary
When a state displays the reverse complement of a sequence, the barcode
is now displayed on the opposite end of the strand

## Implementation Notes
Eg since when modifying the last 3 bases, the first three bases are what
are modified in the reverse complement, so that's
what really refers to the same region of the strand. We discussed
whether the barcode region should instead be removed
entirely (as it probably isn't really the barcode any more in the
reverse complement), but we determined this would
make more sense and allow users to better orient themselves

## Testing
Validated barcode positioning on puzzle with states with and without
reverse complement. Verified it was correctly positioned across
changes in state, target/natural, pip, etc