Fix edge case in custom layout motification detection (#871)
Order of operations issue, caught by newer typescript version in vs code
(and so we now no longer have the ts error being shown in the editor).
As far as I can tell this just meant if moving from no custom layout
being visualized to one with all nulls, we would unnecessarily trigger a
rerender. Which... doesnt really matter. But now we fix the typecheck!