Fix base transforms for puzzles with appended/non-cut oligos (#876)
* In particular, this affects "5P mode" oligos since we previously
assumed oligos were always placed after the strand, but we also had an
off-by-one for 3p mode oligos (as we didn't account for the lack of a
cut point.
* This also fixes previous issues I introduced in pose2d (namely
affecting annotations) where getStrandName/getStrandLabel didn't account
for this either.
* This also updates library selection handling to ensure we're only
selecting within the main sequence region (...again, we hadn't accounted
for 5p oligos) and we now handle the custom numbering transform as part
of our base transforms since otherwise we were doing base transforms on
custom numbering instead of actual base indices.
* This also fixes a skipped library selection (due to it being in a
region without custom numbering) triggering an update (including a new
undo entry) when it should not.
* We also didn't previously handle a null new target oligo order in
transformBaseIndex, which we now handle.